move retry logic to relay coordinator
