+++ crates/coprocessor-node/src/job_executor.rs
@@ -2,29 +2,21 @@
 
 use crate::{
     metrics::Metrics,
-    relayer::{JobRelayer, Relay},
+    relayer::Relay,
     writer::{Write, WriterMsg},
 };
 use alloy::{hex, primitives::Signature, signers::Signer};
 use async_channel::Receiver;
-use ivm_abi::abi_encode_offchain_job_request;
-use ivm_db::{
-    get_all_failed_jobs,
-    tables::{ElfWithMeta, Job, RequestType},
-};
+use ivm_db::tables::{ElfWithMeta, Job, RequestType};
 use ivm_proto::{JobStatus, JobStatusType, VmType};
 use reth_db::Database;
 use std::{
     marker::Send,
     sync::{mpsc::SyncSender, Arc},
-    time::Duration,
 };
-use tokio::{sync::oneshot, task::JoinSet};
+use tokio::task::JoinSet;
 use tracing::{error, info};
 use zkvm_executor::service::ZkvmExecutorService;
-
-/// Delay between retrying failed jobs, in milliseconds.m
-const JOB_RETRY_DELAY_MS: u64 = 250;
 
 /// Errors from job processor
 #[derive(thiserror::Error, Debug)]
@@ -32,21 +24,12 @@ pub enum Error {
     /// database error
     #[error("database error: {0}")]
     Database(#[from] ivm_db::Error),
-    /// ELF with given verifying key already exists in DB
-    #[error("elf with verifying key {0} already exists")]
-    ElfAlreadyExists(String),
     /// Job already exists in DB
     #[error("job already exists")]
     JobAlreadyExists,
     /// Failed to send job to exec queue
     #[error("failed to send to exec queue")]
     ExecQueueSendFailed,
-    /// Could not read ELF from DB
-    #[error("failed reading elf: {0}")]
-    ElfReadFailed(String),
-    /// Could not write ELF to DB
-    #[error("failed writing elf: {0}")]
-    ElfWriteFailed(String),
     /// Invalid VM type
     #[error("invalid VM type")]
     InvalidVmType,
@@ -106,11 +89,10 @@ pub struct JobExecutorConfig {
 pub struct JobExecutor<S, D> {
     db: Arc<D>,
     exec_queue_receiver: Receiver<Job>,
-    job_relayer: Arc<JobRelayer>,
     zk_executor: ZkvmExecutorService<S>,
     task_handles: JoinSet<Result<(), Error>>,
     metrics: Arc<Metrics>,
-    config: JobExecutorConfig,
+    num_workers: usize,
     writer_tx: SyncSender<WriterMsg>,
     relay_tx: tokio::sync::mpsc::Sender<Relay>,
 }
@@ -125,30 +107,27 @@ where
     pub fn new(
         db: Arc<D>,
         exec_queue_receiver: Receiver<Job>,
-        job_relayer: Arc<JobRelayer>,
         zk_executor: ZkvmExecutorService<S>,
         metrics: Arc<Metrics>,
-        config: JobExecutorConfig,
+        num_workers: usize,
         writer_tx: SyncSender<WriterMsg>,
         relay_tx: tokio::sync::mpsc::Sender<Relay>,
     ) -> Self {
         Self {
             db,
             exec_queue_receiver,
-            job_relayer,
             zk_executor,
             task_handles: JoinSet::new(),
             metrics,
-            config,
+            num_workers,
             writer_tx,
             relay_tx,
         }
     }
 
-    /// Starts both the relay retry cron job and the job processor, and spawns `num_workers` worker
-    /// threads
+    /// Spawns `num_workers` worker tasks.
     pub async fn start(&mut self) {
-        for _ in 0..self.config.num_workers {
+        for _ in 0..self.num_workers {
             let exec_queue_receiver = self.exec_queue_receiver.clone();
             let db = Arc::clone(&self.db);
             let zk_executor = self.zk_executor.clone();
@@ -168,16 +147,6 @@ where
                 .await
             });
         }
-
-        let db = Arc::clone(&self.db);
-        let job_relayer = Arc::clone(&self.job_relayer);
-        let metrics = Arc::clone(&self.metrics);
-        let max_retries = self.config.max_retries;
-        let writer_tx = self.writer_tx.clone();
-
-        self.task_handles.spawn(async move {
-            Self::start_job_retry_task(db, job_relayer, metrics, max_retries, writer_tx).await
-        });
     }
 
     /// Start a single queue puller worker task.
@@ -312,134 +281,4 @@ where
             }
         }
     }
-
-    /// Retry jobs that failed to relay
-    async fn start_job_retry_task(
-        db: Arc<D>,
-        job_relayer: Arc<JobRelayer>,
-        metrics: Arc<Metrics>,
-        max_retries: u32,
-        writer_tx: SyncSender<WriterMsg>,
-    ) -> Result<(), Error> {
-        loop {
-            // Jobs that we no longer want to retry
-            let mut jobs_to_delete = Vec::new();
-
-            let retry_jobs = match get_all_failed_jobs(db.clone()) {
-                Ok(jobs) => jobs,
-                Err(e) => {
-                    error!("error retrieving relay error jobs: {:?}", e);
-                    continue;
-                }
-            };
-
-            // Retry each once
-            for mut job in retry_jobs {
-                let id = job.id;
-                let result = match job.request_type {
-                    RequestType::Onchain => {
-                        job_relayer.relay_result_for_onchain_job(job.clone()).await
-                    }
-                    RequestType::Offchain(_) => {
-                        let job_params = (&job).try_into()?;
-                        let job_request_payload = abi_encode_offchain_job_request(job_params);
-                        job_relayer
-                            .relay_result_for_offchain_job(job.clone(), job_request_payload)
-                            .await
-                    }
-                };
-
-                match result {
-                    Ok(receipt) => {
-                        info!("successfully retried job relay for job: {:?}", hex::encode(id));
-                        jobs_to_delete.push(id);
-
-                        // Save the relay tx hash and status to DB
-                        job.relay_tx_hash = receipt.transaction_hash.to_vec();
-                        job.status.status = JobStatusType::Relayed as i32;
-                        job.status.failure_reason = None;
-
-                        // We are not in a rush, so we can wait for the write
-                        let (tx, rx) = oneshot::channel();
-                        writer_tx
-                            .send((Write::JobTable(job.clone()), Some(tx)))
-                            .expect("db writer broken");
-                        let _ = rx.await;
-                    }
-                    Err(e) => {
-                        if job.status.retries == max_retries {
-                            metrics.incr_relay_err(&FailureReason::RelayErrExceedRetry.to_string());
-                            jobs_to_delete.push(id);
-                            info!(
-                                id = hex::encode(id),
-                                "queueing un-broadcastable job for deletion"
-                            );
-                        } else {
-                            error!(
-                                id = hex::encode(id),
-                                ?e,
-                                "report this error: failed to retry relaying job",
-                            );
-                            job.status.retries += 1;
-
-                            // We are not in a rush, so we can wait for the write
-                            let (tx, rx) = oneshot::channel();
-                            writer_tx
-                                .send((Write::FailureJobs(job.clone()), Some(tx)))
-                                .expect("db writer broken");
-                            let _ = rx.await;
-                        }
-                    }
-                }
-
-                // Before retrying another job, wait to reduce general load on the system
-                tokio::time::sleep(Duration::from_millis(JOB_RETRY_DELAY_MS)).await;
-            }
-
-            for job_id in &jobs_to_delete {
-                // We are not in a rush, so we can wait for the write
-                let (tx, rx) = oneshot::channel();
-                writer_tx
-                    .send((Write::FailureJobsDelete(*job_id), Some(tx)))
-                    .expect("db writer broken");
-                let _ = rx.await;
-            }
-            tokio::time::sleep(Duration::from_secs(30)).await;
-        }
-    }
-}
-
-/// Relay job result to the EVM.
-pub async fn relay_job_result(
-    mut job: Job,
-    job_relayer: Arc<JobRelayer>,
-    writer_tx: SyncSender<WriterMsg>,
-) -> Result<(), FailureReason> {
-    let id = job.id;
-
-    let relay_receipt_result = match job.request_type {
-        RequestType::Onchain => job_relayer.relay_result_for_onchain_job(job.clone()).await,
-        RequestType::Offchain(_) => {
-            let job_params = (&job).try_into().map_err(|_| FailureReason::RelayErr)?;
-            let job_request_payload = abi_encode_offchain_job_request(job_params);
-            job_relayer.relay_result_for_offchain_job(job.clone(), job_request_payload).await
-        }
-    };
-
-    let relay_tx_hash = match relay_receipt_result {
-        Ok(receipt) => receipt.transaction_hash,
-        Err(e) => {
-            error!("failed to relay job {:?}: {:?}", id, e);
-            writer_tx.send((Write::FailureJobs(job), None)).expect("db writer broken");
-
-            return Err(FailureReason::RelayErr);
-        }
-    };
-
-    // Save the relay tx hash and status to DB
-    job.relay_tx_hash = relay_tx_hash.to_vec();
-    job.status.status = JobStatusType::Relayed as i32;
-    writer_tx.send((Write::JobTable(job), None)).expect("db writer broken");
-
-    Ok(())
 }
